requirement.txt
absl-py==2.3.1
Adafruit-Blinka==8.66.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-neopixel==6.3.17
adafruit-circuitpython-pixelbuf==2.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
av==12.3.0
Babel==2.10.3
beautifulsoup4==4.11.2
beniget==0.4.1
binho-host-adapter==0.1.6
blinker==1.5
Bottleneck==1.3.5
Brotli==1.0.9
cachetools==5.5.2
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
contourpy==1.0.7
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
et-xmlfile==1.0.1
Flask==2.2.2
flatbuffers==20181003210633
fonttools==4.38.0
fs==2.4.16
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
gpiozero==2.0.1
grpcio==1.74.0
h5py==3.14.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
jdcal==1.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.10.3
keras==2.15.0
kiwisolver==0.0.0
lgpio==0.2.2.0
libarchive-c==2.9
libclang==18.1.1
libevdev==0.5
lxml==4.9.2
lz4==4.0.2+dfsg
Markdown==3.8.2
MarkupSafe==2.1.2
matplotlib==3.6.3
meson==1.5.1
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==0.0.0
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
opencv-python==4.8.1.78
openpyxl==3.0.9
opt_einsum==3.4.0
packaging==23.0
pandas==1.5.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.30
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==10.0.1
pluggy==1.0.0+repack
ply==3.11
protobuf==4.25.8
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.13
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pyftdi==0.57.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyserial==3.5
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-prctl==1.8.1
pythran==0.11.0
pytz==2022.7.1
pyudev==0.24.0
pyusb==1.3.1
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rfc3987==1.3.8
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1a4
rsa==4.9.1
RTIMULib==7.2.1
scipy==1.10.1
SCons==4.4.0
sense-hat==2.6.0
simplejpeg==1.8.1
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
soupsieve==2.3.2
sparkfun_qwiic_as7265x==2.0.0
sparkfun_qwiic_i2c==2.0.0
spidev==3.5
ssh-import-id==5.10
sympy==1.11.1
sysv-ipc==1.1.0
tables==3.7.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
toml==0.10.2
tqdm==4.64.1
twython==3.8.2
typing_extensions==4.4.0
ufoLib2==0.14.0
uritemplate==4.1.1
urllib3==1.26.12
v4l2-python3==0.3.5
videodev2==0.0.4
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.2
zipp==1.0.0
